requirements.txt
Flask==2.2.2
gunicorn==20.1.0
webvtt-py==0.5.1
requests
